requirements.txt
flask
markdown
requests
google-genai
python-dotenv
gunicorn
PyYAML
